# api/openapi.yaml
components:
...
  responses:
    Error:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      description: 'Structured error: 400 invalid input, 401 unauthenticated, 403
        denied, 404 missing or not owned, 409 conflict/busy/stale generation, 410 expired result or operation, 422
        unsupported, 502 runtime failure, 504 timeout.'
  schemas:
...
    Error:
      additionalProperties: false
      properties:
        error:
          additionalProperties: false
          properties:
            code:
              type: string
            message:
              type: string
          required:
          - code
          - message
          type: object
      required:
      - error
      type: object
...
    ExecResult:
      additionalProperties: false
      properties:
        exitCode:
          type: integer
        stderr:
          type: string
        stdout:
          type: string
        truncated:
          type: boolean
      required:
      - stdout
      - stderr
      - exitCode
      type: object
...
    ToolRequest:
      additionalProperties: false
      properties:
        args:
          items:
            type: string
          type: array
        timeoutMs:
          default: 30000
          maximum: 300000
          minimum: 1
          type: integer
      required: []
      type: object
...
info:
  description: Internal-network sandbox infrastructure. Control and service APIs require
    no token or browser authorization. X-Cellbox-Client-ID optionally selects a data
    and idempotency namespace, not an authenticated identity; the default is config.clientId,
    or internal. Single-writer durable state; no automatic command replay or implicit wake.
  title: Cellbox REST API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /v1/boxes/{id}/credentials:
    put:
      summary: Install admitted credential slots in one generation-fenced request
      description: A 204 response acknowledges all writes against the requested generation. A failed batch can have applied some slots; retry the complete batch. Values are base64 strings, each decoding to 1..1048576 bytes, at most 2097152 bytes total.
      operationId: writeCredentialBatch
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [expectedGeneration, slots]
              properties:
                expectedGeneration:
                  type: integer
                  minimum: 1
                slots:
                  type: object
                  minProperties: 1
                  maxProperties: 32
                  propertyNames:
                    pattern: '^[A-Za-z][A-Za-z0-9_]{0,63}$'
                  additionalProperties:
                    type: string
                    contentEncoding: base64
      responses:
        "204":
          description: All slots installed in the requested generation
        default:
          $ref: '#/components/responses/Error'
...
  /v1/boxes/{id}/tools/{tool}:
    post:
      description: Arguments are validated against full-match patterns. Tool definitions
        are server profile configuration; arbitrary debug commands are forbidden.
        This synchronous request is not replay safe.
      operationId: runTool
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      - in: path
        name: tool
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToolRequest'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecResult'
          description: Success
        default:
          $ref: '#/components/responses/Error'
      summary: Run an immutable admitted tool as debug
...
security: []
servers:
- description: Local default; use TLS for remote deployment
  url: http://127.0.0.1:8090